feat: add functionality for compress\decompress; refactor: cp
