+++ src/modules/path/checkPath.mjs
