+++ src/modules/path/getUserData.mjs
