+++ src/modules/path/position.mjs
@@ -1,4 +1,4 @@
-import { txtPos } from './textArgs.mjs';
+import { txtPos } from '../textArgs.mjs';
 
 export default function pos(dir) {
   return `${txtPos} ${dir}`;
